`GET /organizations/:keycloakId` uses permissions
Restrict visibility to the `GET /organizations/:keycloakId` endpoint
using the permission system. A grant of `view` on the organization in
the service is now required (via the `/user..Permission` endpoints) in
order for the caller of the organizations endpoint to see the
organization's service IDs.

To make this practical, a new function `has_data_provider_permission`
patterned after the existing `has_funder_permission` is hereby added.

Issue #1744 Add organizations endpoint