Log a warning on missing `PG...` env vars
Without this change, simply missing some PG environment variables can
be difficult to track down. With this change, a warning is logged when
one of the usual `PG...` environment variables is missing. These are
not directly used by the service but are a convention for underlying
libraries, that is why it is a warning and may not be an error.

Issue #1829 When migrations fail, it is hard to tell why