Include `Opportunity` in `Proposals`
This continues our pattern of including deep related objects in our
entities.  This results in redundancy in some cases, and eventually it
would be useful to have this kind of data returned in the API bundle
rather than as part of the object directly, but it will work for now.

Issue #1850 [Enhancement]: Embed opportunity in response to GET /proposals