Deploy to a fresh, new production environment
With this change, deployment to a new "Production-2025" environment
occurs when the same conditions for a deployment to the existing
"Production" environment occur. The "Production" environment will
become a "Demo" environment soon, after confirming that the new
"Production-2025" environment is in good shape.

Issue #1638 Update our deployment infrastructure to support a "demo"
    server