WIP
This commit does a lot -- I want to break it into two pieces, but
ultimately it makes a new `File` type and migrates our
`presignedPostRequest` handlers to use the file entity instead of taking
parameters.  This allows us to associate S3 file uuids with specific
users who created them, which in turn we'll use for validation of `file`
fieldValues in an upcoming feature.