Add `/files` endpoints
Prior to this commit we had no entity in the PDC to represent files that
users had uploaded to S3.  This needs to change as we begin to use files
in more places.  Having a file object (with a `createdBy`) let's us
track the most simple forms of ownership and eventually we can expand
to using a permission system if necessary.

This entity essentially replaces the concept of
a `PresignedPostRequest`, and creating one will come with a one-time
presigned post that allows the user to provide the actual data
associated with the file to the relevant storage location.

Issue #1689 Support file types for BaseFields