Start versioning beyond version zero
Until recently, the PDC back-end software has largely been in an
experimental or prototype state. At least two parties (besides the
team developing the PDC back-end) call it. We also anticipate real
data to be sent to the PDC soon. Even if the software remains
experimental for a while, the API has matured over the past three-plus
years since its first commit. It is time to call it version one.