Separate unit and int test in CI
CI tests were configured in a way that ran integration setup for unit
tests as well.  This shouldn't have been a problem (outside of wasted
time running unnecessary migrations), but for some reason it was
conflicting with some upcoming unit test mocks.

That should be looked into separately for posterity, but practically
speaking we shouldn't be running that setup for unit tests.

This separates unit test files to use a modified configuration
/ isolates the integration test setup for just integration tests.

Co-authored-by: Harry Minsky <harry.minsky@gmail.com>