Use host networking and uncommon ports
This allows use of the "localhost" hostname throughout, especially
with the `AUTH_SERVER_ISSUER` and `S3_ENDPOINT` variables.

The alternative would be to modify one's hosts file to add service
names so that they work both inside and outside containers. That is
still an option but I'd like the default to simply work out of the
box without any monkeying around.