Remove `unprocessed` s3 key validation checks
This commit removes all references to, and enforcement of,
an `unprocessed/` prefix on our s3 keys. Since we no longer
force that that be the prefix for our keys, we want to make
sure we don't expect it or make any confusing references to it.

Issue #1886 Enforce or remove Unprocessed s3 key checks