Wait more appropriately for auth container
On first run, the auth container takes longer than a minute on my
machine to start, migrate, etc., before it is ready for commands.
When volumes persist the data (second and later runs), the auth
container takes much less than a minute to start, migrate etc. With
this change, the wait is tied to the server actually listening.