Update S3 tooling to support multiple regions
Although we only expect to use a single region we do specify the region
a given file's data is stored in alongside other bucket information.
This means when accessing file data we will want to be able to provide
that region (which in turn may not be the same region as defined via env
vars).

This change allows the creation of a new S3 client for each region
accessed.