Normalize S3 Buckets
We have learned a bit about how DigitalOcean handles S3, and for that
reason we realized that it isn't enough to store a bucket name (and
a region) but we actually need the endpoint of the bucket since DO's S3
does not provide cross-region access to buckets.

We're making an improvement to the way bucket information associated
with a given file's data is stored in the file entity.  This information
will be used soon when accessing the files data / generating presigned
URIs.

Issue #1901 Expand internal S3 patterns to enable multi-region bucket access