Synchronize with the central instance
Until now, we strictly avoided a reference to an instance of the
software herein contained. In this special case we now reference it
because base fields are essential data built in the central instance
of the PDC and are needed for any meaningful development work against
the PDC API. With this change, a developer of a PDC App that calls the
PDC does not need to specify the location of the central instance in a
file in order to use the `compose-ci.yml` file for development and
testing. In the rare case of pulling images and then working offline,
the `PDC_SYNC_URL` can be set back to `http://localhost:3030` to nerf
base field synchronization.

Issue #565 Docker for automated integration tests