Fix broken migration
We had a migration which would fail in the event of any bulk upload task
data. This was due to the introduction of a new `uuid` constraint on
storage keys.

This fix removes that constraint, allowing the old storage keys to map
to the new `files` entity type.

Issue #1921 Migrations fail if bulk upload data existed