Add `attachmentsArchiveFileId` to BulkUploadTask
Our bulk uploads are going to support the use of files for field values.
To do this we need to provide a way for a user to provide those files.
We've decided the best way to do this is to allow the user to upload
a file archive which the bulk upload processor would download, extract,
and convert into specific files.

This commit adds that to the bulk upload task (though does not update
the processing script to use it).

Issue #1858 Augment Bulk Upload to support file BaseField types