Standardize authContext for bulk upload
The auth context of any task (particularly bulk upload task) is
currently fabricated based on whoever created the bulk upload task.

To load the bulk upload task we needed to load as a superuser (which is
OK -- the task runner itself is a superuser), but we want to switch to
the auth context of the bulk uploader as early as possible.  This change
switches to using that context sooner, and also encapsulates the act of
generating that context into its own function.