Use the official node docker image
Bitnami is not going to support all the images they used to anymore.
See https://github.com/bitnami/charts/issues/35164

The plain (Debian) bookworm image was 1.5GiB (very big) and the slim
version of the same image didn't come with `curl` which we use in the
`compose-ci.yml` file to run some scripts. So install `curl` in the
slim image. Also, due to warnings about UIDs under 1000, update the
user ID to 1002 instead of 902.

Issue #1947 Update docker image