Add a GH action to test Docker image on PRs
Without this commit, we don't find out if the `docker buildx` command
succeeds or fails until a merge to main. While it is still possible for
the command to fail on merge to main with this change, we are much more
likely to get earlier feedback (at PR time) as to whether the Docker
image will be successfully built.

Fixes #1949 Include a docker buildx command in PR CI
See also Issue #1947 for more context.