Refactor Docker actions into single `docker.yml`
Before this change, the `docker buildx` command was in the `build.yml`
alongside the usual `npm run build` command (needed for the image).
Also before this change, a new test of the `docker buildx` command was
in a separate file, divorced from the companion build-and-push job.
With this change, the two related jobs are next to one another with
conditions and comments to delineate them.

Issue #1949 Include a docker buildx command in PR CI

See also related issues #1947 and #1950