Fix BulkUpload integration tests
We had tests written in a way where they passed even though they had not
been properly updated to reflect changes to the bulk upload API (for
instance tests like "fails if fileName is invalid" which failed, but not
for that reason.

This fixes those issues while also moving to a pattern that should allow
us to avoid the same mistake in future thanks to more explicit typing
related to the test data.