Rename local file variable
We recently added the concept of `proposalDataFile` to bulk uploads, but
did not update the variable names within the bulk upload processor to
match the new name.

Didn't matter at the time, but now that we're about to incorporate the
`attachmentsArchiveFile` to the bulk upload processing we will benefit
from the more specific local name as well.

Issue #1858 Augment Bulk Upload to support file BaseField types