Use `aws-sdk-client-mock` for s3 testing
We had been manually looking for various network calls from the s3
client library, catching them via `nock`, and returning mocked values.

This was OK for simple get requests, in theory, though even then it was
still brittle (if the S3 client changed behavior internally we would
need to update our mocks to reflect those changes).

We're soon going to need to mock more complex operations -- specifically
`PUT` for uploading data to S3 as part of the new bulk upload processing
functionality.

There is a `aws-sdk-client-mock` library which generally seems to be
recognized as the Right Way to mock aws functionality.

Issue #1994 Move to aws-sdk-client-mock for tests involving S3