Support `file` values in bulk upload
We want users to be able to associate file attachments with their
proposals when performing a bulk upload.  This adds logic to the bulk
upload processor to download file attachments, extract them, and (if
used) register them as files associated with the uploading user.

Issue #1858 Augment Bulk Upload to support file BaseField types