Bulk upload task logs in `GET /tasks/bulkUploads`
When errors are encountered during bulk upload, they are now saved,
and when the bulk upload task is retrieved via API call, the same logs
are presented in a `logs` array. A UI may examine an entry's `isError`
property to know whether the log entry represents an error and may
display the `message` value to the user, which is guaranteed to be
present.

Issue #731 Report Bulk Upload preflight errors via API