Rename `logger` in graphile tasks
The term `logger` is confusing in the context of tasks.  This is
because we have our own pattens for logging in most places, but when
logging for a task a logger is injected by the graphile-worker library.

To help prevent confusion we want to rename that logger to
`graphileLogger` both when calling it and when passing it around to
various functions in the context of the process.

Issue #2017 Rename logger in the contest of graphile jobs