Lock uuid major version
The uuid library dropped support for commonjs.  To migrate to ESM we
need to make changes to our jest setup.  This is possibly worth the
investment since ESM won't be going away, but it's a large enough lift
that we need to be able to prioritize it over time.

Until then we are going to lock into v11