Decorate files with presigned URLs
Files in the PDC cannot be directly accessed since they are on private
buckets.  The way to allow a user to view the contents of a file is to
generate a presigned download URL.

This commit adds logic to decorate objects with pre-signed download urls
on retrieval.

Issue #1912 Decorate file field values with File objects + presigned access URLs