Improve `POST /changemakers` documentation
Use `null` as `POST /changemakers` org ID example.

This is the likely common case, on creation of a changemaker we don't
expect the Keycloak Organization ID to be available. Later we expect
to use `PATCH /changemakers/{id}` to supply it when needed.

The reason to keep the org UUID property in the example is to
highlight that it is possible to specify it. The reason to use `null`
as the first example value is to highlight that it is not required.

Issue #2070 adding new changemakers shouldn't require keycloak id
    (misleading docs)