Create more specific Auth type for operations
Our database operations only require specific aspects of an
authentication context (the identity of the user as well as any roles
the user has).  This is also true for some utility functions that help
extract values from an AuthContext.

We're about to require identity information when loading users so that
we can be controlled about what data to expose depending on that
identity.  This means that we need this more specific type for our
generator functions in order to load the user in the first place.