Update `User` queries to use authContext
User queries prior to this assumed that they were always being called in
a way where the calling user has the right to see everything about that
user.  We're going to want to have the query be more specific / auth
context aware.

More specifically we only want to expose a user's permissions if the
auth context is that user or a system administrator.

This may eventually become more complex (e.g. a user with manage
permissions on a changemaker might be able to see if a user has any
permissions against that changemaker), but we're starting with the more
blunt / conservative approach for now.

Issue #2037 Decorate BulkUploads with full User entities