Afford sensitivity filter on `GET /baseFields`
A necessary part of the change to omit forbidden base fields by
default is to grant the ability to see them when needed. Without this
change, all base fields are always shown, including forbidden ones, on
a `GET /baseFields`. With this change, a new parameter called
`sensitivityClassifications` is available to specify which sensitivity
classifications to show (and not). By default (no parameter specified),
it is as if the parameter is set to `!["forbidden"]`.

Issue #1849 [Enhancement]: Endpoint for baseFields get to allow
    criteria