Single logout via a small Swagger-UI plugin
To allow Keycloak logout from the OpenAPI documentation, this uses an
OpenAPI plugin as a separate module. The reason for clean separation
is inclusion of the "dom" module, which has no place in the node
project except to be added to the `swagger-ui` browser code. Without
this change, the logout only applies to the OpenAPI docs page and the
PDC Keycloak session would remain. With this change, the logout really
logs out of PDC Keycloak entirely.

Issue #1915 Single logout from OpenAPI docs