Enable auto-merge for Dependabot PRs
We *only* want this workflow because we require human approval, which
means all this workflow does is save the reviewer the need to also click
the "merge" button after they click approve.

If we ever remove that approval requirement we should remove this
workflow.  I explored the ability for the workflow to only run if the
requirement exists and unfortunately the github token for actions
doesn't have permission to access that information. I sent a request to
support to add that feature, so maybe some day we can augment the
workflow to be safer.

Issue #2090 Set up dependabot auto-merge