Support partial editing of applicationFormFields
We want to allow a user to update the label and description of a given
applicationFormField.  It's important to note that the "core" values of
an application form field (in particular the base field type)
is not editable.  That could become editable in future, but for now the
"right" way to change the type of data in a field is to create a new
application form field and upload new data against it.

Issue #2065 Support editing of application form fields