Provide undefined vs null signals for some queries
In JavaScript `null` and `undefined` are different things, but when
a value is sent to pg it is converted to a sql `NULL` in either case.
This means if we have a patch query (where a user can provide any subset
of values to update, and the expected behavior is that undefined values
are not updated) we lose the ability to set nullable columns to `NULL`.

This change will allow us to write queries that can determine if the
user was intending to explicitly set a column to `NULL` as opposed to
simply not providing a value (in which case the user is signaling they
do not want to update the value at all)

Issue #1472 Add ability to reset a nullable field back to null