Adjust dependabot cadence
In a perfect world with infinite resources we would want daily
dependabot PRs, but that isn't our world yet.

We are slowing non-security dependabot bumps to monthly; this should not
affect critical patches as those are handled outside of the dependabot
configuration.  We also are adding a 7-day "cooldown" which means
a given release has to be out for 7 days before Dependabot will consider
opening a PR for it.  This has positive security implications because it
means if there are any security issues they're more likely to be found
and addressed before we incorporate the version bump

Issue #2140 Modify dependabot cadences