Use generated id for some tests
This is not a complete fix to a common problem where we use hard coded
values in our tests instead of id values from generated test objects.
The inspiration for making some progress on this front is we had cases
where the hard coded value was of type `string` and only worked because
ajv coercion is disabled.

We're about to disable ajv coercion, which would ultimately break some
of these tests.