Update string validators to expect strings
We are about to disable ajv coercion, which means that field values
(which are always of type string) cannot be validated against json
schemas but instead we need to validate them as strings.

This updates our validators to properly validate strings according to
the relevant expected patterns and other rules.

This means that validating a value will NOT change that value.

Issue #2158 Disable ajv coercion