Disable ajv coercion
We do not want AJV to modify the json values provided by users.  If
a user passes null, that should be treated as null; if a user passes the
string `"42"` that should be treated as a string, not a number.

This makes our API stricter for end users while also making it more
predictable.  It also allows us to set nullable string values to null
instead of empty string.

Issue #2158 Disable ajv coercion