Add test factory for BaseField
We have started to add factory utilities for tests to help avoid lots of
redundant values. This is a factory method for BaseFields which can be
used in upcoming tests; eventually we will want to modify previous tests
to use this utility as well but that's a large enough set of changes
that I'm opting to separate it into a future commit.